### 2025-05-17 #8248 @editorialbot
[PRE REVIEW]: MOSAIK: Multi-Origin Spatial Transcriptomics Analysis and Integration Kit
**Submitting author:** @anthbapt (<a href="http://orcid.org/0000-0002-8514-0250">Anthony Baptista</a>)
**Repository:** https://github.com/anthbapt/MOSAIK
**Branch with paper.md** (empty if default branch): main
**Version:** v0.1
