### 2025-07-15 #8248 @willgearty
@anthbapt could you please suggest 5 reviewers from our [reviewer database](https://reviewers.joss.theoj.org/lookup) or your own ([non-conflicted](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)) extended network? Their github handles are most useful to receive but please don't use "@" to reference them since it will prematurely ping them.
