### 2025-08-13 #8248 @JonnyTran
> :wave: @neuromusic @JonnyTran @haessar @gtonkinhill @jaybee84, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi! I'm not able help review this time.
