### 2025-08-13 #8795 @editorialbot
[REVIEW]: MOSAIK: Multi-Origin Spatial Transcriptomics Analysis and Integration Kit
**Submitting author:** @anthbapt (<a href="http://orcid.org/0000-0002-8514-0250">Anthony Baptista</a>)
**Repository:** https://github.com/anthbapt/MOSAIK
**Branch with paper.md** (empty if default branch): main
**Version:** 1.0.1
