### 2025-08-15 anthbapt/MOSAIK#1 @willgearty
Licenses must be separated for JOSS publication
As part of the review process for your JOSS submission (openjournals/joss-reviews#8795), we've noticed that your `LICENSE.md` does not match an [OSI-approved license](https://opensource.org/licenses/alphabetical). Upon further inspection, I see that you have included an OSI-approved license (MIT) along with a CC license (CC-BY) within this single file. After discussion with some other JOSS editors, we have no problem with you licensing the different components of your package under these different licenses, but they will need to be stored within separate license files for JOSS publication. We recommend that you name them `LICENSE-MIT` and `LICENSE-CCBY`.
