### 2025-09-04 anthbapt/MOSAIK#1 @anthbapt
Hello,

Thank you for the information, I separated my licence into two different files LICENSE-MIT.md and LICENSE-CCBY.md as suggested.

Best wishes,
Anthony
