### 2025-09-24 anthbapt/MOSAIK#5 @haessar
How to use Napari
Following the steps and trying to install the whl, I'm getting dependency errors within the napari IPython console.

I notice that there is a cosmx.yml conda environment file defined in the env/ directory. Therefore, I assume there is a missing installation step. Presumably, you need to outline installation of a conda environment using that environment file (including having conda installed), and then how exactly to install napari with that conda env activated (e.g. https://github.com/napari/napari/tree/v0.4.17?tab=readme-ov-file#installation)
