### 2025-09-24 #8795 @haessar
@willgearty I've had to pause my review as I'm unable to install the software in order to reproduce analysis. I've added some issues which are currently blocking some of the checklist items, including this (https://github.com/anthbapt/MOSAIK/issues/5)
