### 2025-09-25 #8795 @skanwal
Following-up the issue thread to see when the installation [issues](https://github.com/anthbapt/MOSAIK/issues/5) are resolved.
