### 2025-10-20 anthbapt/MOSAIK#5 @haessar
Hi @anthbapt , I'm trying to install the new napari.yml as outlined in the README but getting an error:
`CondaHTTPError: HTTP 503 SERVICE UNAVAILABLE for url <https://conda.anaconda.org/napari/label/bundle_tools/linux-64/repodata.json>
Elapsed: 00:07.865766
CF-RAY: 99181c25bd81cc95-MAN
`

This seems to be related to the use of channel `napari/label/bundle_tools` in the yml. Is there a reason to use this channel beyond the napari packages that are available through `conda-forge`?
