### 2025-11-03 anthbapt/MOSAIK#5 @skanwal
Hi @haessar and @anthbapt - just wanted to let you know that I have tried installing napari using `env/napari.yml` and that has successfully created the envm.
