### 2025-11-04 anthbapt/MOSAIK#5 @skanwal
Hi @haessar - yes, the yml contains this channel:

```
name: napari-test
channels:
  - conda-forge
  - defaults
  - napari/label/bundle_tools
```
