### 2025-11-06 anthbapt/MOSAIK#5 @haessar
@anthbapt, at the step "In the right panel, select the parent folder that contains your CosMx run.", which folder exactly should I be selecting? I tried `flatFiles` but that threw the "not a valid slide" error.
