### 2025-11-06 anthbapt/MOSAIK#5 @anthbapt
Let me know if it is still unclear
