### 2025-11-12 #8795 @haessar
@anthbapt I'm still a little bit confused about how to proceed with reproducing an analysis. Can you clarify for me? Where abouts in the Workflow diagram (Fig 1) am I taken to by running the "Integrating the data step by step"
workflows? I'm assuming from context it's the "SD object .zarr" bubble?

How do the "Python analysis" and subsequent yellow boxes in the Fig 1 workflow correspond to the steps in the readthedocs?

In summary, I'm a little confused about which `tools` scripts I need to run (and in which order) to reproduce the analysis from the paper
