### 2025-11-26 #8795 @haessar
I really value the time it must have taken to put those notebooks together @anthbapt and it really does improve my understanding of the full workflow. I've opened an issue https://github.com/anthbapt/MOSAIK/issues/16 just to address a few teething issues but mostly they all look great.

Just to go back to my previous question, and to help me finally map the workflow to the figure. I assume then that all the steps outlined in the notebooks cover all parts of the figure other than these:

which are the downstream analysis? If so, where does squidpy fit into MOSAIK (I can't see it referenced within the code other than as a xenium.yml dependency), and where do the CompartmentLabels and CellOverlay directories fit in (I also don't see any reference to these in the code or the paper, despite them representing the difference between the two "green circle" sets of analysis input files)?
