### 2025-11-26 anthbapt/MOSAIK#16 @haessar
Notebooks issues
Thanks for adding the notebooks, that makes the overall workflows a lot easier to grasp!

I'll use this issue to tally some of the bugs I encounter / suggestions for improvement while going through it.

- [x]  As the conda env is assumed to be active when running these, it makes sense to add `jupyter` as a dependency in both cosmx.yml and xenium.yml
- [x] There's an inconsistency in section 2 (Setup and Imports). While cosmx.py, _constants.py, sbf.py, etc are required to be in the **working directory** (i.e. the directory where Jupyter is launching the notebooks, which would usually be MOSAIK/notebooks/), the initial validation block is checking that these files exist in the **project directory** (in my case QuarterBrain/). For it to work correctly, I'd need to be running the notebook from within the **project directory** (so that it becomes the **working directory**), and the only way you can really do this in Jupyter (from my experience) is by copying over the notebook into that directory. When the **working directory** and **project directory** are the same, this makes having to define `path` somewhat redundant, but it's probably fine to keep it as is for safety. My recommendation here: add a note or suggestion somewhere (perhaps in README or top of the notebook itself) to make sure the notebook is run from within the **project directory** (i.e. copy it over). This also applies to the Xenium notebook.
- [x] cosmx_QC.py is listed as an optional module, but by the looks of it is never called within the notebook. Correct me if I'm wrong, but the notebook is essentially replicating what cosmx_QC.py would be doing; providing a more interactive path for the analysis/QC. I'd recommend removing it as an optional module (unless my understanding is wrong), and perhaps leaving a note in the notebook and/or README that it is an alternative to running cosmx_QC.py. I assume this also applies to the Xenium notebook and xenium_QC.py.
