### 2025-11-27 anthbapt/MOSAIK#16 @anthbapt
I added the following text just after the section entitled "Integrating the data step by step" in the readme

💡 Alternatively to running `cosmx_QC.py` or `xenium_QC.py` (localised in `src/qc/..`), users interested in a more interactive approach can use the notebooks CosMx_Spatial_Analysis_Workflow.ipynb and Xenium_Spatial_Analysis_Workflow.ipynb (localised in `notebooks/..`) to proceed to the following steps.
