### 2025-11-27 #8795 @anthbapt
I integrated your comments defined in issue #16 and also added this line: "All the tools mentioned in the README can directly load the created zarr object. These include CellTypist (for cell annotation), squidpy (for spatial analysis), and NicheCompass (for niche analysis and cell communication)" in the CosMx and Xenium notebook "Next steps" sections to clarify the connection with the other Python packages. I also added the links to the online squidpy tutorials for both CosMx and Xenium.

Please let me know if anything else needs to be amended.
