### 2025-11-27 #8795 @haessar
Thanks for doing that so quickly @anthbapt. I'm essentially happy to conclude my review, but I'd really value if you could just explain a remaining confusion for me (just for my benefit, it doesn't necessarily require a change to the code/docs): As I mentioned in my previous comment

> where do the CompartmentLabels and CellOverlay directories fit in (I also don't see any reference to these in the code or the paper, despite them representing the difference between the two "green circle" sets of analysis input files)?

I'm still not entirely getting what the differences between these two sets are meant to be in the figure
