### 2025-12-19 #8795 @willgearty
@anthbapt could you please complete the set of additional author tasks listed above?
