### 2026-01-07 #8795 @willgearty
- The title and list of authors for the [Zenodo repository](https://doi.org/10.5281/zenodo.18155537) should match that of the paper. You'll need to go to Zenodo and manually edit the archive metadata.
- Sorry, I meant literal references/citations. Are there citations that you could include for any of the packages that you are already listing in the Related Software section? For example, you should cite [this paper](https://www.nature.com/articles/s41592-021-01358-2) for squidpy.
- Thank you!
