### 2026-01-07 #8795 @anthbapt
Hello @willgearty,

Thank you for your reply. I have edited the metadata of the Zenodo repository, and it should now be aligned. I also added references for the four packages mentioned in the related software section as follows:
“Moreover, well-established libraries such as squidpy [@Palla2022] and scvi-tools [@Gayoso2022] can be easily incorporated into downstream analyses. In addition, we provide seamless integration of Cellpose-SAM [@Pachitariu2025] and Proseg [@Jones2025] for cell segmentation.”

Please let me know if anything should be adjusted.
Thanks for your help!
