### 2026-01-07 #8795 @willgearty
Thanks for adding those citations, @anthbapt 👍

Could you also please add a citation for scanpy (https://doi.org/10.1186/s13059-017-1382-0) and scverse (https://doi.org/10.1038/s41587-023-01733-8)? You should also include the existing Marconato et al citation when referencing the spatialdata package within the Relevant software section.

Once those citations are added the submission should be ready for acceptance!
