### 2026-01-09 #8795 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
