### 2026-01-15 #8795 @anthbapt
Hello @rkurchin ,

 I fixed the mu notation issue by copying unicode character. I checked the generated pdf, it looks fine (see screenshot)

Best wishes,
Anthony
